+++ benchmarks/src/jvmMain/kotlin/com/ae/log/benchmarks/NetworkPipelineBenchmark.kt
@@ -43,9 +43,7 @@ open class NetworkPipelineBenchmark {
 
     @Setup
     fun setup() {
-        AELog.configure {
-            plugin(NetworkPlugin(maxEntries = 200))
-        }
+        AELog.install(NetworkPlugin())
         plugin = AELog.getPlugin<NetworkPlugin>()!!
     }