+++ plugins/network/src/commonTest/kotlin/com/ae/log/network/NetworkRecorderTest.kt
@@ -18,7 +18,7 @@ class NetworkRecorderTest {
 
     @BeforeTest
     fun setUp() {
-        AELog.configure { plugin(NetworkPlugin()) }
+        AELog.install(NetworkPlugin())
         storage = NetworkStorage()
         recorder = NetworkRecorder(storage)
     }